% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdb2igraph.R
\name{cfdb2igraph}
\alias{cfdb2igraph}
\title{Convert cross-feeding database (cfdb) to network (igraph) object}
\usage{
cfdb2igraph(cfdb, mode = "total")
}
\arguments{
\item{cfdb}{A cross-feeding database (cfdb) generated using mdb2cfdb()}

\item{mode}{Whether to calculate forward (first to second), reverse (second to first) or total metabolite exchanges. Default is mode="total".}
}
\description{
Convert cross-feeding database (cfdb) to network (igraph) object
}
\details{
Convert cross-feeding database to network (igraph) object
}
\examples{
cfdb2igraph(allgenomes_cfdb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
