% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdb2pair.R
\name{cfdb2pair}
\alias{cfdb2pair}
\title{Convert cross-feeding database (cfdb) to pairwise matrix}
\usage{
cfdb2pair(cfdb, mode = "total")
}
\arguments{
\item{cfdb}{A cross-feeding database (cfdb) generated using mdb2cfdb()}

\item{mode}{Whether to calculate forward (columns to rows), reverse (rows to columns) or total metabolite exchanges. Default is mode="total".}
}
\description{
Convert cross-feeding database (cfdb) to pairwise matrix
}
\details{
Convert cross-feeding database to pairwise matrix
}
\examples{
cfdb2pair(allgenomes_cfdb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
