% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdb2focal.R
\name{exdb2focal}
\alias{exdb2focal}
\title{Calculation of metabolite exchange potential of each genome}
\usage{
exdb2focal(exdb, exchange = "total", abundance, focal, verbosity = TRUE)
}
\arguments{
\item{exdb}{A exchange database (exdb) generated using mdb2exdb()}

\item{exchange}{Whether to calculate donor, receptor or total metabolite exchange capacities. Default is mode="total".}

\item{abundance}{An optional data frame containing relative abundance data of bacteria}

\item{focal}{An optional focal genome name or vector of genome names to calculate donor potential}

\item{verbosity}{Whether to print the progress of the computation when using abundance data. Default=TRUE}
}
\description{
Calculation of metabolite exchange potential of each genome
}
\details{
Calculation of metabolite exchange potential of each genome
}
\examples{
exdb2focal(allgenomes_exdb)
exdb2focal(allgenomes_exdb, abundance=genome_abundances)
exdb2focal(allgenomes_exdb, abundance=genome_abundances, focal="genome1")
exdb2focal(allgenomes_exdb, focal=c("genome1","genome2"))
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
