% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdb2pair.R
\name{exdb2pair}
\alias{exdb2pair}
\title{Convert exchange database (exdb) to pairwise matrix}
\usage{
exdb2pair(
  exdb,
  exchange = c("forward", "reverse", "total"),
  abundance,
  verbosity = TRUE
)
}
\arguments{
\item{exdb}{An exchange database (exdb) generated using mdb2exdb()}

\item{exchange}{Whether to calculate forward (columns to rows), reverse (rows to columns) or total metabolite exchanges. Default is mode="total".}

\item{abundance}{An optional data frame containing relative abundance data of bacteria}

\item{verbosity}{Whether to print the progress of the computation when using abundance data. Default=TRUE}
}
\description{
Convert exchange database (exdb) to pairwise matrix
}
\details{
Convert exchange database to pairwise matrix
}
\examples{
exdb2pair(allgenomes_exdb)
exdb2pair(allgenomes_exdb, exchange="total")
exdb2pair(allgenomes_exdb, exchange=c("forward","reverse"))
exdb2pair(allgenomes_exdb, exchange="total", abundance=genome_abundances)
exdb2pair(allgenomes_exdb, exchange="total", abundance=genome_abundances, verbosity=FALSE)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
