% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdb2summary.R
\name{exdb2summary}
\alias{exdb2summary}
\title{Summarise pairwise matrix or list of pairwise matrices}
\usage{
exdb2summary(
  exdb,
  summarise = c("genomes", "metabolites", "samples"),
  exchange = "average",
  metabolites,
  genomes,
  samples
)
}
\arguments{
\item{exdb}{A exchange database (exdb) generated using medb2exdb()}

\item{summarise}{Whether to summarise genomes, metabolites or samples. Default is all three.}

\item{exchange}{Whether to calculate genome summaries as average, donor, receptor metabolite exchange capacities. Default is average.}

\item{metabolites}{An optional vector of metabolites for which to run the calculations. Default is all metabolites.}

\item{genomes}{An optional vector of genomes for which to run the calculations. Default is all genomes}

\item{samples}{An optional vector of samples for which to run the calculations. Default is all samples}
}
\description{
Summarise pairwise matrix or list of pairwise matrices
}
\details{
Summarise pairwise matrix or list of pairwise matrices
}
\examples{
exdb2summary(allgenomes_exchange_total)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
