% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedb2exdb.R
\name{gedb2exdb}
\alias{gedb2exdb}
\title{Generate exchange database from a multi-genome metabolite database (gedb)}
\usage{
gedb2exdb(gedb, mode = "strict")
}
\arguments{
\item{gedb}{A metabolite database (gedb) produced by rdb2gedb}

\item{mode}{Whether to calculate strict or loose metabolite exchange capacities. Default is mode="strict".}
}
\description{
Generate exchange database from a multi-genome metabolite database (gedb)
}
\details{
Generate exchange database
}
\examples{
gedb2exdb(allgenomes_gedb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
