% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedb2medb.R
\name{gedb2medb}
\alias{gedb2medb}
\title{Generate a genome database from a reaction database (rdb)}
\usage{
gedb2medb(gedb)
}
\arguments{
\item{rdb}{An rdb object produced by sbml2rdb().}
}
\description{
Classification of metabolites in a single genome into source, transit and sink metabolites
}
\details{
Generate metabolite database
}
\examples{
gedb2medb(allgenomes_gedb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{rdb}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
