% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb2exdb.R
\name{mdb2exdb}
\alias{mdb2exdb}
\title{Generate exchange database from a multi-genome metabolite database (mdb)}
\usage{
mdb2exdb(mdb, mode = "strict")
}
\arguments{
\item{mdb}{A metabolite database (mdb) produced by rdb2mdb or rdbs2mdb}

\item{mode}{Whether to calculate strict or loose metabolite exchange capacities. Default is mode="strict".}
}
\description{
Generate exchange database from a multi-genome metabolite database (mdb)
}
\details{
Generate exchange database
}
\examples{
mdb2exdb(allgenomes_mdb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
