% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medb2exdb.R
\name{medb2exdb}
\alias{medb2exdb}
\title{Generate exchange database from a metabolite database (medb)}
\usage{
medb2exdb(
  medb,
  mode = "strict",
  abundance,
  metabolites,
  genomes,
  samples,
  verbosity = TRUE
)
}
\arguments{
\item{medb}{A metabolite database (medb) produced by gedb2medb}

\item{mode}{Whether to calculate strict or loose metabolite exchange capacities. Default is mode="strict".}

\item{abundance}{An optional dataframe containing relative abundance data of bacteria}

\item{metabolites}{An optional vector of metabolites for which to run the calculations. Default is all metabolites.}

\item{genomes}{An optional vector of genomes for which to run the calculations. Default is all metabolites.}

\item{samples}{An optional vector of samples for which to run the calculations. Default is all samples}

\item{verbosity}{Whether to print the progress of the computation when using abundance data. Default=TRUE}
}
\description{
Generate exchange database from a multi-genome metabolite database (gedb)
}
\details{
Generate exchange database
}
\examples{
medb2exdb(allgenomes_gedb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
