% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadeep.R
\name{metadeep}
\alias{metadeep}
\title{Quantify metabolite exchanges in microbiomes}
\usage{
metadeep(
  sbml,
  mode = "strict",
  abundance,
  summarise = c("genomes", "metabolites", "samples"),
  exchange = "average",
  metabolites,
  genomes,
  samples,
  verbosity = TRUE
)
}
\arguments{
\item{sbml}{An object of class SBMLR, an sbml file or, most commonly, a vector of smbl files.}

\item{summarise}{Whether to summarise genomes, metabolites or samples. Default is all three.}

\item{exchange}{Whether to calculate genome summaries as average, donor, receptor metabolite exchange capacities. Default is average.}

\item{metabolites}{An optional vector of metabolites for which to run the calculations. Default is all metabolites.}

\item{genomes}{An optional vector of genomes for which to run the calculations. Default is all genomes}

\item{samples}{An optional vector of samples for which to run the calculations. Default is all samples}
}
\description{
Quantify metabolite exchanges in microbiome. Note this is a wrapper of many functions in metaDEEP.
 Run the pipeline step by step to achieve more customisation capacity
}
\details{
Quantify metabolite exchanges in microbiomes
}
\examples{
metadeep(sbml_files)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
