% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair2igraph.R
\name{pair2igraph}
\alias{pair2igraph}
\title{Convert a pairwise metabolite exchange matrix into a network (igraph) object}
\usage{
pair2igraph(pair, mode = "undirected")
}
\arguments{
\item{pair}{A pairwise metabolite exchange matrix or a list of pairwise metabolite exchange matrices}

\item{mode}{Whether to generate a directed ("forward" or "reverse") or undirected network. Default is mode="undirected".}
}
\description{
Convert cross-feeding database (cfdb) to network (igraph) object
}
\details{
Convert a pairwise metabolite exchange matrix into a network (igraph) object
}
\examples{
pair2igraph(allgenomes_pair_total)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
