% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair2summary.R
\name{pair2summary}
\alias{pair2summary}
\title{Summarise pairwise matrix or list of pairwise matrices}
\usage{
pair2summary(pair)
}
\arguments{
\item{pair}{A cross-feeding database (cfdb) generated using mdb2cfdb()}
}
\description{
Summarise pairwise matrix or list of pairwise matrices
}
\details{
Summarise pairwise matrix or list of pairwise matrices
}
\examples{
pair2summary(allgenomes_exchange_total)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
