% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb2gedb.R
\name{rdb2gedb}
\alias{rdb2gedb}
\title{Generate a genome database from a reaction database (rdb)}
\usage{
rdb2gedb(rdb)
}
\arguments{
\item{rdb}{An rdb object produced by sbml2rdb().}
}
\description{
Classification of metabolites in a single genome into source, transit and sink metabolites
}
\details{
Generate genome database
}
\examples{
rdb2gedb(genome1_rdb)
sbml2rdb("data/genome1.sbml") \%>\% rdb2gedb()
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{rdb}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
