% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb2igraph.R
\name{rdb2igraph}
\alias{rdb2igraph}
\title{Convert a reaction database into a network (igraph) object}
\usage{
rdb2igraph(rdb)
}
\arguments{
\item{rdb}{A reaction database (rdb object) produced by sbml2rdb().}
}
\description{
Convert cross-feeding database (cfdb) to network (igraph) object
}
\details{
Convert a reaction database into a network (igraph) object
}
\examples{
rdb2igraph(allgenomes_rdb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
