% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbs2mdbs.R
\name{rdbs2mdbs}
\alias{rdbs2mdbs}
\title{Classification of metabolites of multiple genomes}
\usage{
rdbs2mdbs(rdbs)
}
\arguments{
\item{rdb}{An rdb object produced by sbml2rdb().}
}
\description{
Classification of metabolites in multiple genomes into source, transit and sink metabolites
}
\details{
Classification of metabolites
}
\examples{
rdbs2mdbs(allgenomes_rdbs)
list.files(path = "data", pattern = "\\\\.sbml$", full.names = TRUE) \%>\% sbml2rdb() \%>\% rdbs2mdbs()
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{rdb}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
