% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbml2rdb.R
\name{sbml2rdb}
\alias{sbml2rdb}
\title{Convert SBML to reaction database}
\usage{
sbml2rdb(sbml)
}
\arguments{
\item{sbml}{An object of class SBMLR, an sbml file or a vector of smbl files.}
}
\description{
Convert a SBML file into a tibble contain details of all reactions in a genome
}
\details{
Convert SBML to database
}
\examples{
sbml2rdb("data/genome1.sbml")
readSBML("data/genome1.sbml") \%>\% sbml2rdb()
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
