% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmls2rdbs.R
\name{sbmls2rdbs}
\alias{sbmls2rdbs}
\title{Convert multiple SBMLs into a multi-genome reaction database (rdbs)}
\usage{
sbmls2rdbs(sbmls)
}
\arguments{
\item{sbml}{A list of sbml files.}
}
\description{
Convert multiple SBML files into a list of tibbles containing details of all reactions in all genomes
}
\details{
Convert multiple SBMLs to a reaction database
}
\examples{
sbmls2rdbs(c("data/genome1.sbml","data/genome2.sbml"))
list.files(path = "data", pattern = "\\\\.sbml$", full.names = TRUE) \%>\% sbmls2rdb()
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
