% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor.R
\name{donor}
\alias{donor}
\title{Calculation of donor potential}
\usage{
donor(cfdb, abundance, focal)
}
\arguments{
\item{cfdb}{A cross-feeding database (cfdb) generated using mdb2cfdb()}

\item{abundance}{An optional data frame containing relative abundance data of bacteria}

\item{focal}{An optional focal genome name to calculate donor potential}
}
\description{
Calculation of donor potential of each genome
}
\details{
Calculation of donor potential
}
\examples{
donor(allgenomes_cfdb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
