% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdb2pair.R
\name{exdb2pair}
\alias{exdb2pair}
\title{Convert exchange database (exdb) to pairwise matrix}
\usage{
exdb2pair(exdb, metabolites, genomes, samples)
}
\arguments{
\item{exdb}{An exchange database (exdb) generated using medb2exdb()}

\item{metabolites}{An optional vector of metabolites for which to run the calculations. Default is all metabolites.}

\item{genomes}{An optional vector of genomes for which to run the calculations. Default is all metabolites.}

\item{samples}{An optional vector of samples for which to run the calculations. Default is all samples}
}
\description{
Convert exchange database (exdb) to pairwise matrix
}
\details{
Convert exchange database to pairwise matrix
}
\examples{
exdb2pair(allgenomes_exdb)
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
