% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb2mdb.R
\name{rdb2mdb}
\alias{rdb2mdb}
\title{Classification of metabolites of a single genome}
\usage{
rdb2mdb(rdb)
}
\arguments{
\item{rdb}{An rdb or rdbs object produced by sbml2rdb().}
}
\description{
Classification of metabolites in a single genome into source, transit and sink metabolites
}
\details{
Classification of metabolites
}
\examples{
rdb2mdb(genome1_rdb)
sbml2rdb("data/genome1.sbml") \%>\% rdb2mdb()
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{rdb}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
