% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor.R
\name{receptor}
\alias{receptor}
\title{Calculation of receptor potential}
\usage{
receptor(cfdb, abundance, focal)
}
\arguments{
\item{cfdb}{A cross-feeding database (cfdb) generated using mdb2cfdb()}

\item{abundance}{An optional data frame containing relative abundance data of bacteria}

\item{focal}{An optional focal genome name or vector of genome names to calculate receptor potential}
}
\description{
Calculation of receptor potential of each genome
}
\details{
Calculation of receptor potential
}
\examples{
receptor(allgenomes_cfdb)
receptor(allgenomes_cfdb, abundance=genome_abundances)
receptor(allgenomes_cfdb, abundance=genome_abundances, focal="genome1")
receptor(allgenomes_cfdb, focal=c("genome1","genome2"))
}
\references{
Keating, S.M. et al. (2020). SBML Level 3: an extensible format for the exchange and reuse of biological models. Molecular Systems Biology 16: e9110
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{SBML}
\keyword{product}
\keyword{reactant}
\keyword{reaction}
\keyword{tibble}
